Revert [setup.xml/menu.xml] update channelselection entry This reverts commit cffcdf8.
Sorry that I revert this one... as I initially did not want to merge
this. There is maybe more that could be re-assigned to the menu
structure and there are also facturs in this commit that need
discussion. I prefer first to get all structurise and then change it at
Thanks to @Littlesat