[Directories.py] sanitizeFilename, fix blacklist (#4006) Original code comes from here: https://pypi.org/project/sanitize-filename/
But the blacklist has been modified which breaks valid filenames.

This commit returns to the original blacklist, allowing spaces and parenthesis.

The old code also breaks picon names where parenthesis and spaces especially are used. This affects the new UTF8 SNP algorithm where a 1:1 relationship is expected
Thanks to @Huevos