Resolve DVB-T region and epg.dat location for two different server boxes Previously the EPG was also taken from the box where you tried to retrieve the DVB-T region from, but this is not always the main server box where you should take the EPG from.
There are only on two locations one setting retrieved from a server box.
Therefore I do not lead the complete settings anymore, but parse them
and return the value when it is found. I know I parse now when DVB-T and
epg is on the same box two times the same file but it keeps the code
more compact and readable
Thanks to @Littlesat